requirements.txt
transformers
pytorch
numpy